+++ tests/Database/SQLite/DatabaseConnectorTest.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Illuminate\Tests\Database\SQLite;
+
+use Illuminate\Database\Connectors\SQLiteConnector;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+use stdClass;
+
+class DatabaseConnectorTest extends TestCase
+{
+    protected function tearDown(): void
+    {
+        m::close();
+    }
+
+    public function testMemoryDatabasesMayBeConnectedTo()
+    {
+        $dsn = 'sqlite::memory:';
+        $config = ['database' => ':memory:'];
+        $connector = $this->getMockBuilder(SQLiteConnector::class)->onlyMethods(['createConnection', 'getOptions'])->getMock();
+        $connection = m::mock(stdClass::class);
+        $connector->expects($this->once())->method('getOptions')->with($this->equalTo($config))->willReturn(['options']);
+        $connector->expects($this->once())->method('createConnection')->with($this->equalTo($dsn), $this->equalTo($config), $this->equalTo(['options']))->willReturn($connection);
+        $result = $connector->connect($config);
+
+        $this->assertSame($result, $connection);
+    }
+
+    public function testFileDatabasesMayBeConnectedTo()
+    {
+        $dsn = 'sqlite:'.__DIR__;
+        $config = ['database' => __DIR__];
+        $connector = $this->getMockBuilder(SQLiteConnector::class)->onlyMethods(['createConnection', 'getOptions'])->getMock();
+        $connection = m::mock(stdClass::class);
+        $connector->expects($this->once())->method('getOptions')->with($this->equalTo($config))->willReturn(['options']);
+        $connector->expects($this->once())->method('createConnection')->with($this->equalTo($dsn), $this->equalTo($config), $this->equalTo(['options']))->willReturn($connection);
+        $result = $connector->connect($config);
+
+        $this->assertSame($result, $connection);
+    }
+}
